Atualizado metodo delete

+++ src/main/java/com/dio/live/controller/JornadaTrabalhoController.java
@@ -1,6 +1,8 @@
 package com.dio.live.controller;
 
 import com.dio.live.model.JornadaTrabalho;
+import com.dio.live.model.Usuario;
+import com.dio.live.repository.JornadaRepository;
 import com.dio.live.service.JornadaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,12 +11,17 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/jornada")
 public class JornadaTrabalhoController {
+
     @Autowired
     JornadaService jornadaService;
+
+    @Autowired
+    JornadaRepository jornadaRepository;
 
     @PostMapping
     public JornadaTrabalho createJornada(@RequestBody JornadaTrabalho jornadaTrabalho){
@@ -40,11 +47,13 @@ public class JornadaTrabalhoController {
 
     @DeleteMapping("/{idJornada}")
     public ResponseEntity deleteByID(@PathVariable("idJornada") Long idJornada) throws Exception {
-       try {
-           jornadaService.deleteJornada(idJornada);
-       }catch (Exception e){
-           System.out.println(e.getMessage());
-       }
-        return (ResponseEntity<JornadaTrabalho>) ResponseEntity.ok();
+        Optional<JornadaTrabalho> jornadaTrabalho = this.jornadaRepository.findById(idJornada);
+
+        if (jornadaTrabalho.isPresent()) {
+            jornadaService.deleteJornada(idJornada);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
