+++ src/main/java/com/dio/live/controller/UsuarioController.java
@@ -1,7 +1,7 @@
 package com.dio.live.controller;
 
-import com.dio.live.model.JornadaTrabalho;
 import com.dio.live.model.Usuario;
+import com.dio.live.repository.UsuarioRepository;
 import com.dio.live.service.UsuarioService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -9,16 +9,20 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/usuario")
 public class UsuarioController {
 
     @Autowired
+    UsuarioRepository usuarioRepository;
+
+    @Autowired
     UsuarioService usuarioService;
 
     @PostMapping
-    public Usuario createUsuario(@RequestBody Usuario usuario){
+    public Usuario createUsuario(@RequestBody Usuario usuario) {
         return usuarioService.saveUsuario(usuario);
     }
 
@@ -29,21 +33,23 @@ public class UsuarioController {
 
     @GetMapping("/{idUsuario}")
     public ResponseEntity<Usuario> getUsuarioByID(@PathVariable("idUsuario") Long idUsuario) throws Exception {
-        return  ResponseEntity.ok(usuarioService.getById(idUsuario).orElseThrow(() -> new NoSuchElementException("Not found!")));
+        return ResponseEntity.ok(usuarioService.getById(idUsuario).orElseThrow(() -> new NoSuchElementException("Not found!")));
     }
 
     @PutMapping
-    public Usuario updateUsuario(@RequestBody Usuario usuario){
+    public Usuario updateUsuario(@RequestBody Usuario usuario) {
         return usuarioService.updateUsuario(usuario);
     }
 
     @DeleteMapping("/{idUsuario}")
-    public ResponseEntity deleteByID(@PathVariable("idUsuario") Long idUsuario) throws Exception {
-        try {
+    public ResponseEntity<Usuario> deleteByID(@PathVariable("idUsuario") Long idUsuario) {
+        Optional<Usuario> usuario = this.usuarioRepository.findById(idUsuario);
+
+        if (usuario.isPresent()) {
             usuarioService.deleteUsuario(idUsuario);
-        }catch (Exception e){
-            System.out.println(e.getMessage());
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
         }
-        return (ResponseEntity<Usuario>) ResponseEntity.ok();
     }
 }
